pyproject.toml
[project]
name = "llm-engineering"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google>=3.0.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "google-genai>=1.41.0",
    "google-generativeai>=0.8.5",
]
